refactor(gui): extract commands/voice tests to sibling file
Pure verbatim relocation of the `#[cfg(test)] mod tests` block from
`primer-gui/src/commands/voice.rs` into a new sibling
`src/commands/voice/tests.rs`. No behaviour change, no incidental fixes.

`commands/voice.rs` is a non-root module, so its submodule resolves to
`src/commands/voice/tests.rs`. The test module is a descendant, so
`use super::*` still reaches every private parent item — zero visibility
churn. voice.rs drops 841 → 559 lines (−282); tests.rs is 281 lines.

Verified on both feature configs (default + speech): default 205 tests,
speech 253 tests — identical to the pre-move baselines. clippy + fmt clean
on both; workspace suite green (51 ok blocks, 0 failures).
